requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
mysql-connector-python==9.1.0
flask_bcrypt==1.0.1
pymysql==1.1.1
cryptography==44.0.0
PyJWT==2.10.1